+++ app/src/main/java/com/example/spending_control/Valores.kt
@@ -1,5 +1,5 @@
 package com.example.spending_control
 
-class Valores (var idvalor: String, var valor: Int) {
-    constructor() : this("", 0)
+class Valores (var idvalor: String, var valor: Double) {
+    constructor() : this("", 0.0)
 }
